@@ -1,48 +1,41 @@
 {
   "name": "picture_catalog",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.5.2",
-    "cloudinary-react": "^1.6.7",
-    "connected-react-router": "6.5.2",
-    "history": "4.10.1",
-    "merge": "1.2.1",
-    "popper.js": "^1.16.0",
-    "react": "16.11.0",
-    "react-bootstrap": "^1.3.0",
-    "react-bootstrap-icons": "^1.0.0",
-    "react-dom": "16.11.0",
-    "react-redux": "7.1.1",
-    "react-router": "5.1.2",
-    "react-router-dom": "5.1.2",
-    "react-scripts": "^3.4.1",
-    "reactstrap": "8.1.1",
-    "redux": "4.0.4",
-    "redux-thunk": "2.3.0",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.0.1",
-    "svgo": "1.3.0"
+    "@cloudinary/react": "^1.14.1",
+    "@cloudinary/url-gen": "^1.21.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^6.4.8",
+    "@mui/material": "^6.4.8",
+    "bootstrap": "^5.3.3",
+    "react": "^19.0.0",
+    "react-bootstrap": "^2.10.9",
+    "react-bootstrap-icons": "^1.11.5",
+    "react-dom": "^19.0.0",
+    "react-dropzone": "^14.3.8",
+    "react-hook-form": "7.54.2",
+    "react-router": "^7.4.0",
+    "react-router-dom": "^7.4.0",
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
-    "@types/history": "4.7.3",
-    "@types/jest": "24.0.19",
-    "@types/node": "12.11.6",
-    "@types/react": "16.9.9",
-    "@types/react-dom": "16.9.2",
-    "@types/react-redux": "7.1.5",
-    "@types/react-router": "5.1.2",
-    "@types/react-router-dom": "5.1.0",
-    "@types/reactstrap": "8.0.6",
-    "@typescript-eslint/parser": "^2.5.0",
-    "cross-env": "6.0.3",
-    "eslint": "^6.5.1",
-    "eslint-plugin-flowtype": "^3.13.0",
-    "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.16.0",
-    "nan": "^2.14.1",
-    "typescript": "3.6.4"
+    "@testing-library/react": "16.3.0",
+    "@testing-library/jest-dom": "6.6.3",
+    "@eslint/js": "^9.23.0",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.13.11",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
+    "@typescript-eslint/parser": "^8.27.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.56.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-react": "^7.37.4",
+    "globals": "^16.0.0",
+    "nan": "^2.22.2",
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "start": "react-scripts start",
