@@ -1,22 +1,9 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import { MemoryRouter } from 'react-router-dom';
-import App from './App';
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import { App } from './App';
 
-it('renders without crashing', () => {
-    const storeFake = (state: any) => ({
-        default: () => {},
-        subscribe: () => {},
-        dispatch: () => {},
-        getState: () => ({ ...state })
-    });
-    const store = storeFake({}) as any;
-
-    ReactDOM.render(
-        <Provider store={store}>
-            <MemoryRouter>
-                <App/>
-            </MemoryRouter>
-        </Provider>, document.createElement('div'));
+test('renders React working text', () => {
+    render(<App />);
+    const textElement = screen.getByText(/React is working!/i);
+    expect(textElement).toBeInTheDocument();
 });
